requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==41.0.4
distlib==0.3.6
Django==4.2.5
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.0
platformdirs==3.5.1
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.23.0
virtualenvwrapper-win==1.2.7